# api/openapi.yaml
openapi: "3.1.0"
info:
  title: FamilyRoots API
  version: "1.0.0"
  description: |
    ## FamilyRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
...
  schemas:
...
    Error:
      type: object
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri
          example: "https://familyroots.app/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "The request body contains invalid fields."
        instance:
          type: string
          example: "/api/v1/trees/abc/persons"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string
          example: "req_01hzab3k7x"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          example: "sex"
        code:
          type: string
          example: "INVALID_ENUM"
        message:
          type: string
          example: "Must be one of: MALE, FEMALE, OTHER, UNKNOWN"

    # ── Pagination ──────────────────────────────────────────
...
    TokenPair:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT (15 min)
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          example: 900
          description: Access token TTL in seconds
        user:
          $ref: "#/components/schemas/UserProfile"

    # ── User ────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, email, displayName, isVerified, plan, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
          nullable: true
        locale:
          type: string
          example: "en"
        timezone:
          type: string
          example: "America/New_York"
        isVerified:
          type: boolean
        plan:
          type: string
          enum: [FREE, BASIC, PREMIUM, FAMILY, PROFESSIONAL]
        createdAt:
          type: string
          format: date-time
...
  responses:
...
    401Unauthorized:
      description: Unauthorized — missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    422UnprocessableEntity:
      description: Unprocessable Entity — validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    429TooManyRequests:
      description: Too Many Requests — rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# ──────────────────────────────────────────────────────────────
# GLOBAL SECURITY (override per-operation where needed)
# ──────────────────────────────────────────────────────────────
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      operationId: register
      summary: Register a new user account
      description: |
        Creates a new user in the shared `familyroots-system` tenant (configured by
        `DEFAULT_TENANT_SLUG`). No JWT is issued — the account cannot log in until
        the verification email link is clicked.

        **No `tenant_slug` field** — all users join the single shared workspace.
        The first user ever to register becomes `ADMIN`; all subsequent users
        receive the `STANDARD` role.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, given_name, family_name]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                  description: >
                    Must contain at least one uppercase letter and one digit.
                given_name:
                  type: string
                  minLength: 1
                  maxLength: 100
                family_name:
                  type: string
                  minLength: 1
                  maxLength: 100
      responses:
        "204":
          description: Account created — verification email sent. No body, no session cookie.
        "409":
          description: Email already registered in this tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"
        "429":
          $ref: "#/components/responses/429TooManyRequests"
  /auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: Authenticate with email and password
      description: |
        Returns an access token (JSON body) and sets an `HttpOnly` refresh token
        cookie. Login is blocked with `403 account-not-verified` until the
        user's email has been verified (either via the link in the registration
        email or manually by an Admin).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                rememberMe:
                  type: boolean
                  default: false
                  description: Extends refresh token TTL from 30 days to 90 days
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
              description: "HttpOnly SameSite=Lax refresh_token cookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Email not verified (`type` = `account-not-verified`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "423":
          description: Account locked after too many failed attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/429TooManyRequests"
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: Invalidate the current session
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        "204":
          description: Logged out — refresh token revoked and cookie cleared
        "401":
          $ref: "#/components/responses/401Unauthorized"
  /auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshToken
      summary: Issue a new access token using the refresh cookie
      security:
        - CookieAuth: []
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, expiresIn]
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
                    example: 900
        "401":
          description: Refresh token missing, expired, or revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/429TooManyRequests"
  /auth/verify-email:
    post:
      tags: [Authentication]
      operationId: verifyEmail
      summary: Verify email address with token from verification email
      description: |
        Marks the user's email as verified. The token is a 64-character hex
        string embedded in the verification link emailed at registration time.
        After verification the user can log in normally.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: 64-character hex token from the verification email
      responses:
        "204":
          description: Email verified — no body
        "401":
          description: Token invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/resend-verification:
    post:
      tags: [Authentication]
      operationId: resendVerification
      summary: Resend the email verification link
      description: |
        Generates a new verification token and sends a fresh verification email.
        Always returns `204` regardless of whether the email exists or is already
        verified, to prevent account enumeration.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "204":
          description: Verification email sent (or silently ignored if not applicable)
        "429":
          $ref: "#/components/responses/429TooManyRequests"
  /auth/forgot-password:
    post:
      tags: [Authentication]
      operationId: forgotPassword
      summary: Request a password reset email
      description: |
        Sends a password reset link to the given email address.

        - If the email is **unknown** — returns `204` silently (prevents enumeration).
        - If the account **email is not verified** — returns `403 account-not-verified`.
          The user must verify their email before they can reset their password.
          Use `POST /auth/resend-verification` to get a new verification link.
        - If the account is **verified** — sends a reset email and returns `204`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "204":
          description: Reset email sent (or silently ignored for unknown address)
        "403":
          description: |
            Account email not verified (`type` = `account-not-verified`).
            Verify the email first before requesting a password reset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/429TooManyRequests"
  /auth/reset-password:
    post:
      tags: [Authentication]
      operationId: resetPassword
      summary: Reset password using token from reset email
      description: |
        Sets a new password for the account identified by the reset token.
        On success the token is invalidated and all existing refresh tokens
        for the account are revoked (forcing re-login on all devices).

        Also used as the **account activation** flow for admin-created users:
        the activation email contains a link to this endpoint so the user can
        set their own password on first access.

        If the account email was not previously verified, successfully using
        a reset token automatically marks it as verified.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                  description: Token from the password reset or activation email
                new_password:
                  type: string
                  minLength: 8
                  maxLength: 128
                  description: >
                    Must contain at least one uppercase letter and one digit.
      responses:
        "204":
          description: Password updated — all sessions revoked
        "401":
          description: Token invalid or expired (`type` = `token-invalid` or `token-expired`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: New password does not meet strength requirements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...